fix: proxy the iq request through server for user privacy
