+++ website/src/lib/components/self/EmailViewer.svelte
@@ -450,8 +450,7 @@
 
 	async function getUserIQ(emailAddress: string): Promise<number | null> {
 		try {
-			const [username, instance] = emailAddress.split('#');
-			const response = await fetch(`https://${instance}/api/users/${username}/iq`);
+			const response = await fetch(`/api/users/${emailAddress}/iq`);
 			if (response.ok) {
 				const data = await response.json();
 				return data.iq;
